Unify the TrackScheme Focus, Selection and Navigator handling.
They are all in the TrackScheme navigator class now, that
merged with the now removed SelectionBehaviours class.
