+++ src/main/java/net/trackmate/revised/trackscheme/display/TrackSchemeFrame.java
@@ -13,12 +13,6 @@ import javax.swing.JPanel;
 import javax.swing.SwingUtilities;
 import javax.swing.WindowConstants;
 
-import org.scijava.ui.behaviour.MouseAndKeyHandler;
-import org.scijava.ui.behaviour.io.InputTriggerConfig;
-
-import bdv.BehaviourTransformEventHandler;
-import bdv.viewer.InputActionBindings;
-import bdv.viewer.TriggerBehaviourBindings;
 import net.imglib2.ui.TransformEventHandler;
 import net.imglib2.ui.util.GuiUtil;
 import net.trackmate.revised.context.ContextChooser;
@@ -32,6 +26,13 @@ import net.trackmate.revised.ui.grouping.GroupHandle;
 import net.trackmate.revised.ui.grouping.GroupLocksPanel;
 import net.trackmate.revised.ui.util.InvokeOnEDT;
 import net.trackmate.undo.UndoPointMarker;
+
+import org.scijava.ui.behaviour.MouseAndKeyHandler;
+import org.scijava.ui.behaviour.io.InputTriggerConfig;
+
+import bdv.BehaviourTransformEventHandler;
+import bdv.viewer.InputActionBindings;
+import bdv.viewer.TriggerBehaviourBindings;
 
 public class TrackSchemeFrame extends JFrame
 {
@@ -126,7 +127,7 @@ public class TrackSchemeFrame extends JFrame
 
 		final InputTriggerConfig inputConf = getKeyConfig( optional );
 		trackschemePanel.getNavigator().installActionBindings( keybindings, inputConf );
-		trackschemePanel.getSelectionBehaviours().installBehaviourBindings( triggerbindings, inputConf );
+		trackschemePanel.getNavigator().installBehaviourBindings( triggerbindings, inputConf );
 
 		editFocusVertex = new EditFocusVertexBehaviour( focus, graph, undoPointMarker, trackschemePanel.getDisplay() );
 		trackschemePanel.getDisplay().addTransformListener( editFocusVertex );
